tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-properties.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.oncurrentchange = t.step_func_done(e => {
    assert_equals(e.constructor, AppHistoryCurrentChangeEvent);
    assert_false(e.bubbles);
    assert_false(e.cancelable);
    assert_true(e.isTrusted);
  });
  location.href = "#1";
}, "AppHistoryCurrentChangeEvent's properties");
</script>
